+++ PaginatedFilteredProducts.Infrastructure/Products/Data/ProductsDbContextInitialiser.cs
@@ -2,7 +2,6 @@ using Bogus;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using PaginatedFilteredProducts.Domain.Products.Aggregates;
-using PaginatedFilteredProducts.Domain.Products.ValueObjects;
 
 namespace PaginatedFilteredProducts.Infrastructure.Products.Data;
 
@@ -46,31 +45,24 @@ public class ProductsDbContextInitialiser
 
     public async Task TrySeedAsync()
     {
-        // Default data
-        // Seed, if necessary
         if (!_context.Products.Any())
         {
             var products = new List<Product>();
 
-            // Define available currencies
             var currencies = new[] { "USD", "EUR" };
 
-            // Set up Faker to generate parameter values
             var faker = new Faker();
 
             for (int i = 1; i <= 300; i++)
             {
-                // Generate parameter values
                 var name = faker.Commerce.ProductName();
-                var amount = faker.Random.Decimal(50, 150); // Example: random amount between 50 and 150
+                var amount = faker.Random.Double(50, 150);
                 var currency = faker.PickRandom(currencies);
                 var description = faker.Commerce.ProductDescription();
 
-                // Use the factory method to create a Product instance
                 var product = ProductFactory.CreateProduct(name, amount, currency, description);
 
-                // Add some reviews for each product using a similar approach
-                for (int j = 1; j <= 5; j++) // Example: 5 reviews per product
+                for (int j = 1; j <= 5; j++)
                 {
                     var reviewText = faker.Lorem.Sentence();
                     var reviewRating = faker.Random.Int(1, 5);
